refactor(training): modular architecture + IOC + --opt8bit
Phase 1: Bug fixes
- F1: trigger_neurogenesis preserves ALL state fields
- F2: consistent exam_failures naming (was consecutive_failures)
- F3: preserve best_val_loss/epochs_without_improvement after neurogenesis
- F4: add .float() to net2wider_model glyph_table conversion
- F5: add bitsandbytes as optional dependency
- F6: implement --opt8bit flag for 8-bit AdamW optimizer states

Phase 2-3: Module extraction
- tokenization.py: generate_question_variations, tokenize, format_and_tokenize_dialogue
- stage_config.py: get_stage_config, get_next_dim, get_stage_info
- partitioner.py: partition_corpus_by_mlu, compile_stage_dataset
- exam_compiler.py: compile_exam_sequences_for_age
- corpus.py: compute_corpus_hash, load/save_tokenized_cache
- state_manager.py: plan_exam_failure, run_samantha_eval, trigger_neurogenesis

Phase 4-5: IOC pattern
- strategy.py: TrainingStrategy Protocol + 4 implementations
  - FP32Strategy, BF16Strategy, Opt8BitStrategy, BF16Opt8BitStrategy
  - select_strategy() factory for flag-based selection
- neurogenesis uses strategy.create_optimizer() instead of inline code

Phase 6: Orchestrator
- orchestrator.py: SchoolOrchestrator class (WIP, delegates to monolith for now)

Phase 7: Tests (170 passing)
- test_strategy.py: 11 tests for strategy selection and execution
- test_tokenization.py: 10 tests for tokenization functions
- test_stage_config.py: 10 tests for stage config and progression

Phase 8: Benchmark
- benchmark_opt8bit.py: compares fp32, opt8bit, bf16, bf16+opt8bit
- Results: bf16+opt8bit fastest (0.019s/epoch), bf16 saves ~100MB VRAM

+++ src/bitnet/growth/net2net.py
@@ -216,7 +216,7 @@ def net2wider_model(
 	device = next(old_model.parameters()).device
 
 	if use_glyphs:
-		glyph_table = old_model.glyph_embedding.glyph_table.cpu().numpy()
+		glyph_table = old_model.glyph_embedding.glyph_table.cpu().float().numpy()
 		vocab_embeddings = None
 	else:
 		glyph_table = None
